Add URI parser for omics resource URIs
