+++ omics/transfer/__init__.py
@@ -1,8 +1,9 @@
-from enum import Enum
-from typing import IO, Any, List, Type, Union
+from typing import IO, Any, List, Union
 
 from s3transfer.futures import TransferFuture
 from s3transfer.subscribers import BaseSubscriber
+
+from omics.common.omics_file_types import ExtendedEnum, OmicsFileType
 
 
 class OmicsTransferSubscriber(BaseSubscriber):
@@ -13,64 +14,11 @@ class OmicsTransferFuture(TransferFuture):
     """Future for getting the result of Omics data transfer."""
 
 
-class ExtendedEnum(Enum):
-    """Enum subclass that includes helper methods."""
-
-    @classmethod
-    def list(cls: Type[Enum]) -> List[str]:
-        """Return the list of allowed values in an Enum."""
-        return list(map(lambda c: c.value, cls))  # type: ignore
-
-    @classmethod
-    def from_object(cls: Type[Enum], object: Any) -> Enum:
-        """Convert an object to the Enum type if possible.
-
-        Since Python allows strings to be passed as parameters even when an Enum
-        type is specified, this method converts a string to an enum if it
-        matches one of the allowed values.
-
-        Args:
-            object: the object to convert to an Enum.
-        """
-        if type(object) == cls:
-            return object
-        if type(object) == str:
-            for enum in cls:
-                if object.upper() == enum.value:
-                    return enum
-            valid_values = ", ".join(cls.list())
-            raise AttributeError(f"{cls.__name__} must be one of {valid_values}")
-
-        raise AttributeError(f"Unsupported type for {cls.__name__}: {type(object)}")
-
-
-class ReadSetFileName(ExtendedEnum):
-    """Available read set file names."""
-
-    SOURCE1 = "SOURCE1"
-    SOURCE2 = "SOURCE2"
-    INDEX = "INDEX"
-
-
-class ReferenceFileName(ExtendedEnum):
-    """Available reference file names."""
-
-    SOURCE = "SOURCE"
-    INDEX = "INDEX"
-
-
 class FileTransferDirection(ExtendedEnum):
     """Available transfer directions (UP = upload, DOWN = download)."""
 
     UP = "UP"
     DOWN = "DOWN"
-
-
-class OmicsFileType(ExtendedEnum):
-    """Available file types."""
-
-    READ_SET = "READ_SET"
-    REFERENCE = "REFERENCE"
 
 
 class FileTransfer:
