+++ tests/uriparse/__init__.py
@@ -1,2 +1,6 @@
-VALID_READSET_URI = "omics://123412341234.storage.us-east-1.amazonaws.com/5432154321/readSet/5346184667/source1"
-VALID_REFERENCE_URI = "omics://123412341234.storage.us-east-1.amazonaws.com/5432154321/reference/5346184667/source"
+VALID_READSET_URI = (
+    "omics://123412341234.storage.us-east-1.amazonaws.com/5432154321/readSet/5346184667/source2"
+)
+VALID_REFERENCE_URI = (
+    "omics://123412341234.storage.us-east-1.amazonaws.com/5432154321/reference/5346184667/index"
+)
